requirements.txt
asgiref==3.5.2
Django==4.1.2
psycopg2>=2.8
django-filter==22.1
django-widget-tweaks==1.4.12
mysql-connector-python==8.0.31
mysqlclient==2.1.1
protobuf==3.20.1
sqlparse==0.4.3
tzdata==2022.5
